Man fuck these damn starter scripts bro
why are there 2 separate movement scripts with one hidden in the animator? whose idea was that? how does that benefit anyone? i dont even get how it works or what kind of script that was but it did the same shit the other one did but worse. its gone now.